Done for Now
Finished the design of album, artist, and player pages for mobile devices. It looks good on phones and tablets. Did not create a version for desktops as this is supposed to be an for smaller devices. Some ideas for making this a larger more in depth project would be to add javascript to make the search option work. There could also be some sort of mp3 API that could be linked where music can actually be played as well. The list icon can be designed with functionality still too. I was able to learn a lot from this project and gather some great templates to model other projects after.